gjort småändringar och skrivit så gott som klart passiv ruta

+++ src/Varelser.java
@@ -2,9 +2,9 @@ import java.util.ArrayList;
 
 public class Varelser {
     public String namn;
-    public Integer maxHp;
-    public Integer hp;
-    public Integer styrka;
-    public Integer hastighet;
+    public double maxHp;
+    public double hp;
+    public double styrka;
+    public int hastighet;
     public ArrayList<Void> Attacklista;
 }
